refactor: 배달 요청 데이터 수정
- 요청데이터에 주문자의 id값 제거(주문 id를 통해 구할 수 있다)
