+++ src/main/java/com/sjincho/hun/delivery/dto/DeliveryRequest.java
@@ -1,4 +1,4 @@
-package com.sjincho.hun.delivery.controller;
+package com.sjincho.hun.delivery.dto;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
@@ -11,12 +11,7 @@ public class DeliveryRequest {
     @PositiveOrZero(message = "id 는 음수일 수 없습니다.")
     private final Long orderId;
 
-    @NotNull(message = "주문자 id를 확인해주세요. null 일 수 없습니다.")
-    @PositiveOrZero(message = "id 는 음수일 수 없습니다.")
-    private final Long receiverId;
-
-    public DeliveryRequest(final Long orderId, final Long receiverId) {
+    public DeliveryRequest(final Long orderId) {
         this.orderId = orderId;
-        this.receiverId = receiverId;
     }
 }
