+++ src/main/java/com/sjincho/hun/delivery/service/DeliveryService.java
@@ -1,6 +1,6 @@
 package com.sjincho.hun.delivery.service;
 
-import com.sjincho.hun.delivery.controller.DeliveryRequest;
+import com.sjincho.hun.delivery.dto.DeliveryRequest;
 import com.sjincho.hun.delivery.domain.Delivery;
 import com.sjincho.hun.delivery.dto.DeliveryResponse;
 import com.sjincho.hun.delivery.exception.DeliveryAlreadyRegisterException;
@@ -51,15 +51,14 @@ public class DeliveryService {
         Order order = orderRepository.findById(request.getOrderId()).orElseThrow(() ->
                 new OrderNotFoundException(OrderErrorCode.NOT_FOUND, request.getOrderId()));
 
-        Member receiver = memberRepository.findById(request.getReceiverId()).orElseThrow(() ->
-                new MemberNotFoundException(MemberErrorCode.NOT_FOUND, request.getReceiverId()));
+
+        Member receiver = memberRepository.findById(order.getOrderer().getMemberId()).orElseThrow(() ->
+                new MemberNotFoundException(MemberErrorCode.NOT_FOUND, order.getOrderer().getMemberId()));
 
         deliveryRepository.findByOrderId(request.getOrderId())
                 .ifPresent(delivery -> {
                     throw new DeliveryAlreadyRegisterException(DeliveryErrorCode.NOT_READY_REGISTERED, request.getOrderId());
                 });
-
-        // TODO : order의 주문자가 receiver가 맞는지 검증해야함.
 
         Delivery delivery = Delivery.create(order, receiver);
